base classed screens, added -Wall -Werror

+++ src/game/Game.cpp
@@ -3,7 +3,7 @@
 #include "GameData.h"
 
 GameResult Game::run_game() {
-    for(int i=0;i<context.assetLoader.get_levels().size();i++)
+    for(size_t i=0;i<context.assetLoader.get_levels().size();i++)
     {
         switch (run_level(i)) {
             case EXITED:
